requirements.txt
streamlit
pandas
numpy
scikit-learn
statsmodels
lifelines
bcrypt
cryptography
python-docx
reportlab
SQLAlchemy